mqtt_gateway/requirements.txt
nameko==2.12.0
paho-mqtt==1.5.0